refactor(spark): consolidate HoodieFileScanRDD into hudi-spark-common
All six SparkXXHoodieFileScanRDD classes were textually identical: the
FileScanRDD constructor delta they guarded against (SPARK-37273) only
existed for Spark <= 3.2, and every parameter added since 3.4 has a
default value, so the same 5-arg super call compiles against Spark 3.3
through 4.2. Replace them with a single org.apache.hudi.HoodieFileScanRDD
in hudi-spark-common and move the createHoodieFileScanRDD override from
the six version adapters into BaseSpark3Adapter/BaseSpark4Adapter.
